wangyujiao/ship_classfication_explain/requirements.txt
captum==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
idna==3.4
imageio==2.28.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lime==0.2.0.1
lit==16.0.2
loguru==0.7.2
MarkupSafe==2.1.2
matplotlib==3.5.0
minio==7.1.16
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
setuptools-scm==7.1.0
six==1.16.0
sklearn==0.0.post4
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.186
urllib3==1.26.15
zipp==3.15.0
